Complete ThreadController with dummy views + other minor changes
 - Modify ReplyController and UserController to accommodate changes
 - Modify categories table name column to be unique

+++ app/Http/Controllers/ThreadController.php
@@ -7,17 +7,33 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use DB;
+use Auth;
 
 class ThreadController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
+     * @param  string  $category
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index($category = 'new')
     {
-        //
+        if ($category === 'hot') {
+            $threads = DB::table('threads')->orderBy('upvote', 'desc')->orderBy('updated_at', 'desc')->get();
+        }
+        else if ($category === 'new') {
+            $threads = DB::table('threads')->orderBy('updated_at', 'desc')->get();
+        }
+        else {
+            $category_id = DB::table('categories')->where('name', $category)->first()->id;
+            $threads = DB::table('threads')->where('category_id', $category_id)->get();
+        }
+        
+        return view('show_threads', [
+            'threads' => $threads,
+            'category' => $category,
+        ]);
     }
 
     /**
@@ -27,7 +43,11 @@ class ThreadController extends Controller
      */
     public function create()
     {
-        //
+        $categories = DB::table('categories')->get();
+
+        return view('create_thread', [
+            'categories' => $categories,
+        ]);
     }
 
     /**
@@ -38,22 +58,32 @@ class ThreadController extends Controller
      */
     public function store(Request $request)
     {
+        $categoriesCount = DB::table('categories')->count();
+        $this->validate($request, [
+            'category_id' => 'between:1,'.$categoriesCount,
+            'title' => 'string',
+            'content' => 'string',
+            'tags' => 'string',
+            'sticky' => 'boolean',
+        ]);
+
+        $author_id = Auth::user()->id;
         $category_id = $request->input('category_id');
-        $author_id = $request->input('author_id');
         $title = $request->input('title');
         $sticky = $request->input('sticky');
         $tags = $request->input('tags');
-        $content = $request->$file('content');
-
+        $content = $request->input('content');
 
         DB::table('threads')->insert([
             'title' => $title, 
             'sticky' => $sticky,
-            'content' =>  $content,
-            'author_id' =>  $author_id,
-            'category_id' =>  $category_id,
-            'tags' =>  $tags,
+            'content' => $content,
+            'author_id' => $author_id,
+            'category_id' => $category_id,
+            'tags' => $tags,
         ]);
+
+        return redirect('/home');
     }
 
     /**
@@ -65,6 +95,9 @@ class ThreadController extends Controller
     public function show($id)
     {
         $thread = DB::table('threads')->where('id', $id)->first();
+        if ($thread == null) {
+            abort(404);
+        }
         $replies = DB::table('replies')->where('thread_id', $thread->id)->orderBy('position', 'asc')->get();
 
         return view('show_thread', [
@@ -81,7 +114,20 @@ class ThreadController extends Controller
      */
     public function edit($id)
     {
-        //
+        $thread = DB::table('threads')->where('id', $id)->first();
+        if ($thread == null) {
+            abort(404);
+        }
+        if ($thread->author_id != Auth::user()->id) {
+            abort(401);
+        }
+        
+        $categories = DB::table('categories')->get();
+
+        return view('edit_thread', [
+            'thread' => $thread,
+            'categories' => $categories,
+        ]);
     }
 
     /**
@@ -93,26 +139,43 @@ class ThreadController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        $thread = DB::table('threads')->where('id', $id)->first();
+        if ($thread == null) {
+            abort(401);
+        }
+
+        $author_id = $thread->author_id;
+        if ($author_id != Auth::user()->id) {
+            abort(401);
+        }
+
+        $categoriesCount = DB::table('categories')->count();
+
+        $this->validate($request, [
+            'category_id' => 'between:1,'.$categoriesCount,
+            'title' => 'string',
+            'content' => 'string',
+            'tags' => 'string',
+            'sticky' => 'boolean',
+        ]);
+
         $category_id = $request->input('category_id');
-        $author_id = $request->input('author_id');
         $title = $request->input('title');
         $sticky = $request->input('sticky');
         $tags = $request->input('tags');
-        $content = $request->$file('content');
-
+        $content = $request->input('content');
 
         DB::table('threads')
-
             ->where('id', $id)
             ->update([
-            'title' => $title, 
-            'sticky' => $sticky,
-            'content' =>  $content,
-            'author_id' =>  $author_id,
-            'category_id' =>  $category_id,
-            'tags' =>  $tags
-        ]);
+                'title' => $title, 
+                'sticky' => $sticky,
+                'content' => $content,
+                'category_id' => $category_id,
+                'tags' => $tags,
+            ]);
+
+        return redirect('/home');
     }
 
     /**
@@ -123,8 +186,19 @@ class ThreadController extends Controller
      */
     public function destroy($id)
     {
-        //
+        $thread = DB::table('threads')->where('id', $id)->first();
+        if ($thread == null) {
+            abort(401);
+        }
+
+        $author_id = $thread->author_id;
+        if ($author_id != Auth::user()->id) {
+            abort(401);
+        }
+
+        DB::table('replies')->where('thread_id', $id)->delete();
         DB::table('threads')->where('id', $id)->delete();
 
+        return redirect('/home');
     }
 }
